feat: Transaction errors

+++ providers-api/src/main/kotlin/antlapit/near/api/providers/model/account/CallResult.kt
@@ -5,7 +5,7 @@ import antlapit.near.api.providers.primitives.BlockHeight
 import antlapit.near.api.providers.primitives.CryptoHash
 
 data class CallResult(
-    val result: List<Int> = emptyList(),
+    val result: List<Byte> = emptyList(),
     val logs: List<String> = emptyList(),
     override val blockHeight: BlockHeight,
     override val blockHash: CryptoHash
